+++ server/src/api/middleware/passport/naverStrategy.ts
@@ -2,7 +2,8 @@ import { Request } from "express";
 import passport, { Profile } from "passport";
 import { Strategy as NaverStrategy } from "passport-naver";
 import { CONNECTED_ALREADY, EMAIL_IN_USE, SNSID_IN_USE } from "src/api/message";
-import * as UserService from "src/services/user";
+import * as UserService from "src/services/users";
+const AuthService = UserService.AuthService;
 import { ProfileParser } from "./profileParser";
 
 const sns = "naver";
@@ -65,7 +66,7 @@ const naver = (client: { ID: string; SECRET: string; callbackURL: string }) => {
           /* if user is logged in - connect */
           const user = req.user!;
 
-          if (UserService.checkSnsIdActive(user, sns)) {
+          if (AuthService.checkSnsIdActive(user, sns)) {
             const err = new Error(CONNECTED_ALREADY);
             return done(err, null, null);
           }
@@ -79,7 +80,7 @@ const naver = (client: { ID: string; SECRET: string; callbackURL: string }) => {
             return done(err, null, null);
           }
 
-          await UserService.updateSnsId(user, sns, profile.id);
+          await AuthService.updateSnsId(user, sns, profile.id);
           return done(null, user, "connect");
         } catch (error: any) {
           done(error);
